import partial working - still need contacts import

+++ app/models/account.php
@@ -19,19 +19,45 @@ class Account extends AppModel {
      * @return 
      * @access 
      */
-    function update($identity_id, $data) {
-        # remove old account data
-        $this->deleteByIdentityId($identity_id);
-        
-        # add the new data
-        foreach($data as $account_info) {
-            $account = array('Account' => $account_info);
-            $account['Account']['identity_id'] = $identity_id;
-            $saveable = array('identity_id', 'service_id', 'service_type_id', 'username', 
-                              'account_url', 'feed_url', 'created', 'modified');
-            $this->create();
-            $this->save($account, true, $saveable);
+    function update($identity_id, $data, $replace = false) {
+        if($replace) {
+            # remove old account data
+            $this->deleteByIdentityId($identity_id);
         }
+    
+        # add the new data
+        foreach($data as $item) {
+            # check, if we already have it. we need to do this, even
+            # when not replacing. it could be, that an account is
+            # more than once in the array.
+            $urls = array('Account.account_url' => $item['account_url']);
+            if($item['feed_url']) {
+                # only add test for feed URL, when there actually is one.
+                # else, all the contact things like AIM, Jabber would fail.
+                $urls['Account.feed_url'] = $item['feed_url'];
+            }
+            $conditions = array(
+                'Account.identity_id' => $identity_id,
+                array('OR' => $urls)
+            );
+            $this->recursive = 0;
+            $this->expects('Account');
+            if($this->findCount($conditions) == 0) {
+                $item['identity_id'] = $identity_id;
+                $saveable = array(
+                    'identity_id', 'service_id', 'service_type_id', 'title',
+                    'username', 'account_url', 'feed_url', 'created', 'modified'
+                );
+                $this->create();
+                $this->save($item, true, $saveable);
+            }
+        }
+        
+        return true;
+    }
+        
+    function replace($identity_id, $data) {
+        return $this->update($identity_id, $data, true);
     }
     
     /**
@@ -60,7 +86,6 @@ class Account extends AppModel {
         $accounts = array();
         foreach($data as $item) {
             $account = $item['Account'];
-            $account['key'] = md5($account['id']);
             unset($account['id']);
             unset($account['identity_id']);
             unset($account['created']);
@@ -68,5 +93,9 @@ class Account extends AppModel {
             $accounts[] = $account;
         }
         return $accounts;
+    }
+    
+    public function import($identity_id, $data) {
+        return $this->update($identity_id, $data);
     }
 }
